Disabled x64 support by default

+++ packages/stanza/src/stanza/data/sequence.py
@@ -1,4 +1,4 @@
-from stanza.data import Data, PyTreeData
+from stanza.data import Data, PyTreeData, idx_dtype
 from stanza.dataclasses import dataclass, field, replace
 
 from typing import Any, Generic, TypeVar
@@ -117,7 +117,7 @@ class SequenceData(Generic[T,I]):
             t_off[idx:idx+n_chunks] = infos.start_idx[i] + np.arange(n_chunks) * chunk_stride
             i_off[idx:idx+n_chunks] = i
             idx += n_chunks
-        t_off, i_off = jnp.array(t_off, dtype=jnp.uint64), jnp.array(i_off, dtype=jnp.uint64)
+        t_off, i_off = jnp.array(t_off, dtype=idx_dtype), jnp.array(i_off, dtype=idx_dtype)
 
         return ChunkData(
             elements=self.elements,
@@ -130,9 +130,9 @@ class SequenceData(Generic[T,I]):
     def from_trajectory(elements: Data[T], info: I = None) -> "SequenceData[T,I]":
         info = SequenceInfo(
             info=info,
-            start_idx=jnp.array(0, dtype=jnp.uint64),
-            end_idx=jnp.array(len(elements), dtype=jnp.uint64),
-            length=jnp.array(len(elements), dtype=jnp.uint64)
+            start_idx=jnp.array(0, dtype=idx_dtype),
+            end_idx=jnp.array(len(elements), dtype=idx_dtype),
+            length=jnp.array(len(elements), dtype=idx_dtype)
         )
         sequences = PyTreeData(jax.tree_util.tree_map(lambda x: x[None,...], info))
         return SequenceData(
